# fix performance issue in interpolation