--- a/interpolate_test.go
+++ b/interpolate_test.go
@@ -10,7 +10,10 @@ func TestInterpolate(t *testing.T) {
 	vars := map[string]string{
 		"age": "35",
 	}
-	actual, err := interpolate("${name=Bobby Brown} is ${age} years old", vars)
+	updatedVars := map[string]string{
+		"height": "180",
+	}
+	actual, err := interpolate("${name=Bobby Brown} is ${age} years old and ${height} cm tall", vars, updatedVars)
 	require.NoError(t, err)
-	require.Equal(t, `Bobby Brown is 35 years old`, actual)
+	require.Equal(t, `Bobby Brown is 35 years old and 180 cm tall`, actual)
 }