@WillAbides 2020-05-15
While I was writing up the PR, I realized I could further reduce the instances where `Vars.CloneWith()` is called from interpolate vars.  Now instead of every time enum or default is updated, it requires help to have a variable to interpolate as well.