@alecthomas 2020-05-15
Very nice, thanks Will. I'm happy to leave benchmarks in.